develop(MainActivity):optimize codes by page
